+++ packages/babel-preset-app-react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @inventorjs/babel-preset-app-react
 
+## 5.0.5
+
+### Patch Changes
+
+- change template
+
 ## 5.0.4
 
 ### Patch Changes