+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @inventorjs/core
 
+## 5.0.5
+
+### Patch Changes
+
+- change template
+
 ## 5.0.4
 
 ### Patch Changes